requirements.txt
crewai~=0.41.1
crewai-tools~=0.4.26
langchain-community~=0.2.11
python-dotenv~=1.0.1
duckduckgo-search~=6.2.6
uvicorn~=0.30.1
fastapi~=0.111.1
pydantic~=2.8.2
pandas~=2.2.2
requests~=2.32.3
hubspot-api-client~=9.0.0
